mysql 5.7 implementation, cleanup

+++ Server/go/Skill.hpp
@@ -23,7 +23,7 @@ struct Skill
 {
 	string name;
 	float level;
-	float experience;
+	double experience;
 };
 
 #endif /* SKILL_HPP_ */
